:sparkles: feat:优化文章管理相关接口

+++ src/modules/article/category/article-category.controller.ts
@@ -3,10 +3,14 @@ import { ArticleCategoryService } from "./article-category.service";
 import { ArticleCategoryDto } from "./article-category.dto";
 import { ResponseDto } from "@/common/dto/response.dto";
 import { LoginGuard } from "@/common/guard/login.guard";
+import { ArticleService } from "@/modules/article/article/article.service";
 
 @Controller("category")
 export class ArticleCategoryController {
-    constructor(private readonly articleCategoryService: ArticleCategoryService) {}
+    constructor(
+        private readonly articleCategoryService: ArticleCategoryService,
+        private readonly articleService: ArticleService
+    ) {}
 
     /*
      * 新增文章分类
@@ -30,6 +34,10 @@ export class ArticleCategoryController {
         // 检查是否存在parentId为要删除的id的子分类
         const children = await this.articleCategoryService.findByParentId(id);
         if (children.length > 0) return ResponseDto.error("存在子分类，无法删除");
+
+        // 检查该分类下是否存在文章
+        const articles = await this.articleService.queryByCategoryId(id);
+        if (articles.length > 0) return ResponseDto.error("该分类下存在文章，无法删除");
 
         await this.articleCategoryService.remove(id);
         return ResponseDto.success("删除成功");
